wallet backup warning toast in action, added positive balance check
